Adding CMake interprocedural optimization policy setting
This policy setting will silence a warning when using with a CMake's
interprocedural optimization setting (CMAKE_INTERPROCEDURAL_OPTIMIZATION).

The warning can be silence with a clear conscience as interprocedural
optimization (INTERPROCEDURAL_OPTIMIZATION) is not manually enabled on
any gtest targets and thus maybe relying on the old behaviour to only
enable this on Intel compilers on Linux.

Due to the forced `cmake_minimum_version`, policy settings in CMakeLists
calling this one (including the main CMakeLists) are lost, forcing the
change to be made here.