Fix gmock_gen.py generation of template type names
When an interface includes template types, they get automatically
assigned a fixed name, i.e. T0, T1, T2 etc. This works fine as
long as the types are not used in the signature of any of the
functions that need to be mocked.

Note this fix would not have been necessary if support for Python2
would be dropped. In Python 3.6 and later dictionaries are ordered
by default.

Fixes #3172