pkg-config: fix gtest.pc/gmock.pc library names for the debug build
In case CMAKE_BUILD_TYPE is set to Debug the gtest library is
name libgtestd.a but the link command returned from gtest.pc is
'-lgtest' (without the debug d) and so the linking of dependent
packages fails (see [1] for a buildroot failure example).

Enhance the gtest.pc ang gmock.pc generation to honour the debug 'd'.

[1] http://lists.busybox.net/pipermail/buildroot/2021-October/626382.html

Signed-off-by: Peter Seiderer <ps.report@gmx.net>