Add ScopedEnumPrinter
Scoped enums ("enum class", added in C++11) are not implicitly convertible to
bool, so they are not handled by the existing ConvertibleToIntegerPrinter. This
means that they are printed using the RawBytesPrinter, unless there's a custom
printer for the particular type. Printing them as ints is a little nicer than
the raw bytes.