Add GTEST_FALLBACK_PRINTER
This makes it possible to specify a custom fallback printer that is used instead
of the default raw bytes printer.

One especially useful way to use this is to define it to a symbol that doesn't
exist (e.g. "#define GTEST_FALLBACK_PRINTER MISSING_PRINTER") to enforce
printability of all types that are used in gtest checks at compile time.