Add public interface for XmlUnitTestResultPrinter
Allowing custom gtest_main implementations to instantiate the
XmlUnitTestResultPrinter with either a given file path or any other kind
of std::ostream to write to.

This is useful for e.g. embedded cases where an XML report is still
wanted, but not file system is available, by instantiating with a
std::stringstream and delivering the data via any custom mean.

Related to #1930