Eliminate -Wreturn-type (at least on LCC)
There are four GetName<T>() functions that are expected to have
a limited set of possible <T> types. But compiler does not know about
that, and may complain (and in LCC case, does) about no return
operator in the end of these functions.

This commit adds dummy return to each of these functions to eliminate
this warning.