Avod exit-time destructors in macros when possible
This change attemps to reduce the introduction of exit-time destructors
by gtest/gmock whenever it is possible, to just skip the destructor
during termination. This is useful for codebases that aim to comply with
clang's `-Wexit-time-destructors`. However, this change does not attempt
to make all of gtest/gmock complaint with this warning, but limits
itself to what merely affects user code.

Bug: https://github.com/google/googletest/issues/4803